requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic[email]==2.4.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1